moved settings into seperate file
